test: consolidate duplicated result-test helpers into test/helpers.py
Move the copy-pasted unwrap_hv, inner_element and run_cfg_with helpers
into a shared module, and add a BandResult negative test pinning that a
non-scalar (vector) result is rejected rather than silently plotted.
