+++ test/test_box_whisker_result.py
@@ -15,6 +15,7 @@ from bencher.results.bench_result_base import ReduceType
 from bencher.results.holoview_results.distribution_result.box_whisker_result import (
     BoxWhiskerResult,
 )
+from test.helpers import inner_element as _inner_element
 
 
 class DistBench(bn.ParametrizedSweep):
@@ -60,13 +61,6 @@ def _run_sweep(worker_cls, input_vars, repeats):
         run_cfg=run_cfg,
         plot_callbacks=False,
     )
-
-
-def _inner_element(overlay):
-    """The plot methods return an hv.Overlay wrapping the distribution element."""
-    items = list(overlay)
-    assert len(items) == 1
-    return items[0]
 
 
 class TestBoxWhiskerResult(unittest.TestCase):
