+++ test/test_violin_result.py
@@ -9,6 +9,7 @@ import panel as pn
 import bencher as bn
 from bencher.results.bench_result_base import ReduceType
 from bencher.results.holoview_results.distribution_result.violin_result import ViolinResult
+from test.helpers import inner_element as _inner_element
 
 
 class ViolinBench(bn.ParametrizedSweep):
@@ -45,13 +46,6 @@ def _run_sweep(worker_cls, input_vars, repeats):
         run_cfg=run_cfg,
         plot_callbacks=False,
     )
-
-
-def _inner_element(overlay):
-    """The plot methods return an hv.Overlay wrapping the distribution element."""
-    items = list(overlay)
-    assert len(items) == 1
-    return items[0]
 
 
 class TestViolinResult(unittest.TestCase):
